Horizontal calendar fix - Main Pull Request for Sprint-01  (#25)
* Updated .md file to describe frontend calendar html files. Also added a couple comments on file location TODO work

* Days of the week implemented without data injected from backend

* Add date to each DOTW header

 We are having some problems with the angular to display the return value of the getDay() function.

* Complete getDay function
I think when we add the ability to change the schedule to future weeks we might need to change the day param of getDay, thinking about how we would scale it to accomodate future weeks, might need a second param to add to the current date (ex if someone was jumping to next week it would be newDay.setDate(today.getDate() + 7)).

* Organize events into boxes

Very much a work in progress but the basic outline is there. Better formatting and styling is badly needed but I'm in a rush

* Improve Formatting

Relevant information is extracted and properly placed in OH info boxes in a clean style. Button functionality isn't there and the office hours aren't mapped to the correct day yet.

* Turn OH Event box into widget

It is a somewhat confusing import/export method for widgets. However thisOfficeHoursEventCardCopmonent widget is now easily replicable  for additional usage.

* Allow option for TA to include name in OH card

* Add week selection button and functionality
I still need to add some formatting, doesn't look like how we had it in the figma, shouldn't be too hard to reformat though.

* Format dropdown menu
Added a little extra formatting for the choose week dropdown menu, needs to be touched up more though.

* Fix all models to include hostsYou may have to delete and rebuild the database - see database.md

* Start debugging name displaySee the console.log() function I call from widget.html

* Add SCSS class to hide incorrect OH for each day

Big commit.

* Make OH appear on the correct day of week

This was a really good fix, but there are some small problems - OH card goes on week - empty tds (table datas) still take up space

* Add specification md file - needs work

Each answer needs to be buffed up a good bit, but if we all contribute this won't take long

* Resolve day specification on events and remove table

* add mat grid list and move widget to correct dir and update import paths

* Fix styling on calendar, dates in better positions

* Updated spec doc

* Calendar events display top down

---------

Co-authored-by: ian.bracken@duke.edu <idb15@duke.edu>
Co-authored-by: Christopher Saez <csaez@unc.edu>
Co-authored-by: ethanbyrd03 <ethbyr@ad.unc.edu>