29 connect hosts field to the database (#39)
* Adjust buttons and week selector, remove json file

* Create stack_flow.md

Built upon Chris Saez's highly useful stack diagram and updated it to our current knowledge. This new editable format will be useful as we continue to make discoveries and additions to the codebase.

* Buttons change based on usertype, renamed file to .css in widget

* Add delete functionality to the calendar for non-student users

* Add functionality to edit/update an OH event from the calendar

* Update delete functionality to work for all 5 days of the week

* Fix minor problems in stack_flow.md

* Add more-info modal as widget using MatDialog

Event model needs to be given to the widget to display OH details. Pop-up functionality works and the widget is included in the module.ts

* RSVP Instantiation

Switched the button name from book to RSVP. Adjust paths so that RSVP populates in database upon instantiation now but nothing beyond that. RSVP button is not functional, some typescript work will be needed. Also it is not displayed on the Tinstructor's end yet.

* Add demo data for cloud app and localhost                                              /workspace
\
\ The reset_demo script puts the data into the SQLAlchemy session. The next Mon-Fri from when you run the script will have OH.

* Finish modal

Modal now displays all OH info. Used MAT_DIALOG_DATA injection token instead of @Input because @Input isn't supported MatDialog.

* Add rsvp to other backend services

After refreshing the database, OH events can be created with the rsvp column. RSVP number is displayed for non-students.

* Fix OH demo data and ensure it works fine

* RSVP Functionality

RSVP Button is now functional for students and the updated values are displayed on the Instructor's side as well. Some new functions and routes had to be written within both the frontend and backend . There is not yet a limit on the number of times a student can RSVP which should probably be put into place before the end of sprint 2

* Make calendar fit onto normal sized window

* Extreme commit

Host name now appears on current office hours under 'happening now'

* Disable RSVP after clicked and fix history tab

After refreshing, the user can RSVP again. This is because the component refreshes when the webpage does. The solution is to store the user's id and check every time a click is made.

* Clean up size and styling of displayed events

* Add test to backend/test/services/office_hours/office_hours_test.py

Tests increment_rsvp() within backend/services/office_hours/office_hours.py

* documentation improvement

* Update page paginator size from 25 to 50

* important security boost

* Fix failing test and ensure 100% coverage

One test was failing because a SQLAlchemy error was getting raised due to an OH event in office_hours_data.py getting assigned an id that already existed.

* Update technical spec

---------

Co-authored-by: rmt4897 <rmt4897@live.unc.edu>
Co-authored-by: ian.bracken@duke.edu <idb15@duke.edu>