Fix demo data (#42)
There was a problem with datetime.now(), see my comment. I added pytz as a dependency in the requirements.txt file, and it's working when I rebuilt the container (I think I did this right). This library generates the correct east coast time zone automatically accounting for daylight savings time.